fix(fv): getConditions 4-tuple in Coverage.t.sol + assertEq uint256 cast (fix #3)
CI run #3 errors:
1. test/unit/FounderVesting/Coverage.t.sol 6 destructures of getConditions() used
   3-tuple; new signature returns 4 (ethOverride added). Added empty 4th slot in
   each (L93, L104, L115, L125, L135, L146).

2. EdgeCases.t.sol:779 `assertEq(uint256 delta, 8_000_000 * 1e18 / 3)` was
   ambiguous because `1e18` is a rational/int_const literal. Cast to uint256
   explicitly to disambiguate from int256 overload.

+++ test/unit/immutables/FounderVestingV2EdgeCases.t.sol
@@ -776,7 +776,7 @@ contract FounderVestingV2EdgeCases is Test {
         uint256 delta = lumina.balanceOf(recipient) - before_;
         // TRANCHE_AMOUNT = 8M*1e18 / 3 = 2_666_666_666666666666666666 (integer-truncated)
         assertEq(delta, TRANCHE_AMOUNT);
-        assertEq(delta, 8_000_000 * 1e18 / 3);
+        assertEq(delta, uint256(8_000_000 * 1e18 / 3));
     }
 
     function test_Cobro_Tranche1_FVBalanceDecreasesByTrancheAmount() public {
